% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_microhaps.R
\name{add_covered_by_target_to_features}
\alias{add_covered_by_target_to_features}
\title{Add \code{covered_by_target} listing panel targets that fully contain each feature}
\usage{
add_covered_by_target_to_features(feature_tab, ref_bed_tab)
}
\arguments{
\item{feature_tab}{Feature BED table.}

\item{ref_bed_tab}{Panel location table.}
}
\value{
\code{feature_tab} with \code{covered_by_target} (\code{"uncovered"} if none).
}
\description{
Add \code{covered_by_target} listing panel targets that fully contain each feature
}
\keyword{internal}
