% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_allele_frequency_naive.R
\name{af_parse_aa_calls}
\alias{af_parse_aa_calls}
\title{Load amino acid calls for naive allele frequency estimation}
\usage{
af_parse_aa_calls(path)
}
\arguments{
\item{path}{Path to a TSV of amino acid calls.}
}
\value{
A tibble with columns \code{specimen_name}, \code{target_name}, \code{reads}, and
\code{variant}.
}
\description{
Load amino acid calls for naive allele frequency estimation
}
\keyword{internal}
