% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_allele_frequency_naive.R
\name{af_parse_mh_calls}
\alias{af_parse_mh_calls}
\title{Load microhaplotype calls for naive allele frequency estimation}
\usage{
af_parse_mh_calls(path)
}
\arguments{
\item{path}{Path to a TSV of microhaplotype genotypes.}
}
\value{
A tibble with columns including \code{specimen_name}, \code{target_name},
\code{reads}, and \code{variant}.
}
\description{
Load microhaplotype calls for naive allele frequency estimation
}
\keyword{internal}
