% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqEstimationModel_wrapper.R
\name{bin2STAVE}
\alias{bin2STAVE}
\title{Convert a FEM binary haplotype string to a STAVE variant}
\usage{
bin2STAVE(chars, names, alt_alleles, monos)
}
\description{
Convert a FEM binary haplotype string to a STAVE variant
}
\keyword{internal}
