% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_allele_frequency_naive.R
\name{calculate_af_presence_absence}
\alias{calculate_af_presence_absence}
\title{Calculate allele frequency from presence/absence}
\usage{
calculate_af_presence_absence(allele_table)
}
\arguments{
\item{allele_table}{Allele table with \code{target_name} and \code{variant}.}
}
\value{
A tibble with \code{target_name}, \code{variant}, \code{allele_total},
\code{allele_count}, and \code{freq}.
}
\description{
Calculate allele frequency from presence/absence
}
\keyword{internal}
