% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_allele_frequency_naive.R
\name{calculate_af_read_count_prop}
\alias{calculate_af_read_count_prop}
\title{Calculate allele frequency from within-sample allele proportions}
\usage{
calculate_af_read_count_prop(allele_table)
}
\arguments{
\item{allele_table}{Allele table with \code{specimen_name}, \code{target_name},
\code{variant}, and \code{reads}.}
}
\value{
A tibble with \code{target_name}, \code{variant}, and \code{freq}.
}
\description{
Calculate allele frequency from within-sample allele proportions
}
\keyword{internal}
