% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqEstimationModel_wrapper.R
\name{calculate_avg_COI}
\alias{calculate_avg_COI}
\title{Average COI from a table path or a numeric value}
\usage{
calculate_avg_COI(coi_path)
}
\description{
Average COI from a table path or a numeric value
}
\keyword{internal}
