% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_prevfreq_naive.R
\name{calculate_multilocus_af_prev_presence_absence}
\alias{calculate_multilocus_af_prev_presence_absence}
\title{Calculate multilocus prev/freq by presence/absence}
\usage{
calculate_multilocus_af_prev_presence_absence(multilocus_calls)
}
\arguments{
\item{multilocus_calls}{Table with \code{specimen_name}, \code{variant}.}
}
\value{
A tibble with \code{prev} and \code{freq}.
}
\description{
Calculate multilocus prev/freq by presence/absence
}
\keyword{internal}
