% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_prevfreq_naive.R
\name{calculate_multilocus_af_prev_wsaf_prop}
\alias{calculate_multilocus_af_prev_wsaf_prop}
\title{Calculate multilocus prev/freq weighted by WSAF}
\usage{
calculate_multilocus_af_prev_wsaf_prop(multilocus_calls)
}
\arguments{
\item{multilocus_calls}{Table with \code{specimen_name}, \code{variant}, \code{wsaf}.}
}
\value{
A tibble with \code{prev} and \code{freq}.
}
\description{
Calculate multilocus prev/freq weighted by WSAF
}
\keyword{internal}
