% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_allele_prevalence_naive.R
\name{calculate_prevalence}
\alias{calculate_prevalence}
\title{Calculate allele prevalence}
\usage{
calculate_prevalence(allele_table)
}
\arguments{
\item{allele_table}{Allele table with \code{specimen_name}, \code{target_name}, and
\code{variant}.}
}
\value{
A tibble with \code{target_name}, \code{variant}, \code{prev}, \code{sample_count}, and
\code{sample_total}.
}
\description{
Calculate allele prevalence
}
\keyword{internal}
