% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_prevfreq_naive_variantstring.R
\name{calculate_variant_prevalence}
\alias{calculate_variant_prevalence}
\title{Prevalence and frequency of target variants vs a dataset}
\usage{
calculate_variant_prevalence(target_variants, comparison_variants)
}
\arguments{
\item{target_variants}{Target variant strings.}

\item{comparison_variants}{Dataset variant strings.}
}
\value{
A data frame with \code{variant}, \code{prev}, \code{freq}, \code{sample_total}.
}
\description{
Prevalence and frequency of target variants vs a dataset
}
\keyword{internal}
