% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coi_naive.R
\name{check_coi_format}
\alias{check_coi_format}
\title{Check that an object matches the expected COI estimate format}
\usage{
check_coi_format(df_coi)
}
\arguments{
\item{df_coi}{A data frame of COI estimates.}
}
\value{
Invisibly returns \code{TRUE} if valid.
}
\description{
Check that an object matches the expected COI estimate format
}
\keyword{internal}
