% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqEstimationModel_wrapper.R
\name{check_fem_biallelic}
\alias{check_fem_biallelic}
\title{Keep bi-allelic (and report mono-allelic) amino-acid targets}
\usage{
check_fem_biallelic(input_data)
}
\description{
Keep bi-allelic (and report mono-allelic) amino-acid targets
}
\keyword{internal}
