% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_microhaps.R
\name{check_microhap_alignment_pkgs}
\alias{check_microhap_alignment_pkgs}
\title{Require Biostrings and pwalign for microhaplotype overlap alignment}
\usage{
check_microhap_alignment_pkgs(reason)
}
\arguments{
\item{reason}{Passed to \code{\link[=check_suggested_pkg]{check_suggested_pkg()}}.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Require Biostrings and pwalign for microhaplotype overlap alignment
}
\keyword{internal}
