% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-external.R
\name{check_suggested_pkgs}
\alias{check_suggested_pkgs}
\title{Check several suggested packages}
\usage{
check_suggested_pkgs(pkgs, reason = NULL)
}
\arguments{
\item{pkgs}{Character vector of package names.}

\item{reason}{Optional short description of why they are needed.}
}
\value{
Invisibly returns \code{TRUE} if all packages are available.
}
\description{
Check several suggested packages
}
\keyword{internal}
