% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-external.R
\name{check_variantstring_v1}
\alias{check_variantstring_v1}
\title{Require variantstring 1.x}
\usage{
check_variantstring_v1(reason = NULL)
}
\arguments{
\item{reason}{Passed to \code{\link[=check_suggested_pkg]{check_suggested_pkg()}}.}
}
\value{
Invisibly returns \code{TRUE} if a 1.x version is installed.
}
\description{
Require variantstring 1.x
}
\keyword{internal}
