% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THEREALMcCOIL_wrapper.R
\name{clean_up_mccoil}
\alias{clean_up_mccoil}
\title{Remove McCOIL intermediate files from a working directory}
\usage{
clean_up_mccoil(work_dir)
}
\arguments{
\item{work_dir}{Directory that may contain \verb{McCOIL*} traces.}
}
\description{
Remove McCOIL intermediate files from a working directory
}
\keyword{internal}
