% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_loci_of_interest.R
\name{collapse_aa_allele_table}
\alias{collapse_aa_allele_table}
\title{Collapse overlapping-target amino-acid calls}
\usage{
collapse_aa_allele_table(
  allele_table_to_filter,
  collapse_calls_by_summing = FALSE
)
}
\arguments{
\item{allele_table_to_filter}{Translated calls joined to allele counts.}

\item{collapse_calls_by_summing}{If \code{TRUE}, sum reads across targets.}
}
\value{
Collapsed tibble.
}
\description{
Collapse overlapping-target amino-acid calls
}
\keyword{internal}
