% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileup_specific_snps.R
\name{collapse_snp_allele_table}
\alias{collapse_snp_allele_table}
\title{Collapse overlapping-target SNP calls by summing or picking the best target}
\usage{
collapse_snp_allele_table(
  allele_table_to_collapse,
  collapse_calls_by_summing = FALSE
)
}
\arguments{
\item{allele_table_to_collapse}{SNP calls joined to allele counts.}

\item{collapse_calls_by_summing}{If \code{TRUE}, sum reads across targets.}
}
\value{
Collapsed tibble.
}
\description{
Collapse overlapping-target SNP calls by summing or picking the best target
}
\keyword{internal}
