% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_prevfreq_naive_variantstring.R
\name{compute_prevfreq_for_group}
\alias{compute_prevfreq_for_group}
\title{Compute prev/freq for one loci group via variantstring}
\usage{
compute_prevfreq_for_group(group_loci, aa_table)
}
\arguments{
\item{group_loci}{Tibble with \code{gene_id} and \code{aa_position}.}

\item{aa_table}{Amino acid calls in long form.}
}
\value{
Tibble with \code{variant}, \code{prev}, \code{freq}, \code{sample_total}.
}
\description{
Compute prev/freq for one loci group via variantstring
}
\keyword{internal}
