% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcifer_ibd_wrapper.R
\name{create_dcifer_allele_freq_input}
\alias{create_dcifer_allele_freq_input}
\title{Read allele frequencies into Dcifer list format}
\usage{
create_dcifer_allele_freq_input(
  allele_freq_path,
  allele_list,
  target_name_col = "target_name",
  target_value_col = "seq"
)
}
\description{
Read allele frequencies into Dcifer list format
}
\keyword{internal}
