% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcifer_slaf_wrapper.R
\name{create_dcifer_allele_table_input}
\alias{create_dcifer_allele_table_input}
\title{Read an allele table TSV for Dcifer wrappers}
\usage{
create_dcifer_allele_table_input(
  allele_table_path,
  specimen_name_col = "specimen_name",
  target_name_col = "target_name",
  target_value_col = "seq"
)
}
\description{
Read an allele table TSV for Dcifer wrappers
}
\keyword{internal}
