% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcifer_slaf_wrapper.R
\name{create_dcifer_coi_input}
\alias{create_dcifer_coi_input}
\title{Read a COI table into a named vector aligned with a Dcifer allele list}
\usage{
create_dcifer_coi_input(
  coi_path,
  allele_list,
  specimen_name_col = "specimen_name"
)
}
\description{
Read a COI table into a named vector aligned with a Dcifer allele list
}
\keyword{internal}
