% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcifer_ibd_wrapper.R
\name{create_dcifer_specimen_metadata_input}
\alias{create_dcifer_specimen_metadata_input}
\title{Read specimen metadata for population-specific Dcifer runs}
\usage{
create_dcifer_specimen_metadata_input(
  specimen_metadata_path,
  coi,
  specimen_name_col = "specimen_name",
  pop_name_col = "population"
)
}
\description{
Read specimen metadata for population-specific Dcifer runs
}
\keyword{internal}
