% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqEstimationModel_wrapper.R
\name{create_FEM_input}
\alias{create_FEM_input}
\title{Build the genotype matrix and related FEM inputs for one group}
\usage{
create_FEM_input(input_data, groups, group_id)
}
\description{
Build the genotype matrix and related FEM inputs for one group
}
\keyword{internal}
