% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLociBiallelicModel_wrapper.R
\name{create_MultiLociBiallelicModel_input}
\alias{create_MultiLociBiallelicModel_input}
\title{Create Multi-Loci Biallelic Model input}
\usage{
create_MultiLociBiallelicModel_input(
  input_path,
  loci_group,
  aa_sample_occurence_cut_off = 0
)
}
\arguments{
\item{input_path}{Path to amino-acid calls TSV.}

\item{loci_group}{Path to loci-group TSV (\code{group_id}, \code{gene_id}, \code{aa_position}).}

\item{aa_sample_occurence_cut_off}{Amino-acid calls must occur in more than
this number of samples to be included.}
}
\value{
A list (\code{MLBM_object}) with \code{MLBM_data}, \code{staves_data},
\code{by_group_table}, and \code{groups}.
}
\description{
Create Multi-Loci Biallelic Model input
}
\keyword{internal}
