% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpslice_wrapper.R
\name{create_snpslice_allele_table_input}
\alias{create_snpslice_allele_table_input}
\title{Read an allele table with counts for SNP-Slice}
\usage{
create_snpslice_allele_table_input(
  allele_table_path,
  specimen_name_col = "specimen_name",
  target_name_col = "target_name",
  target_value_col = "aa",
  target_count_col = "reads"
)
}
\description{
Read an allele table with counts for SNP-Slice
}
\keyword{internal}
