% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcifer_ibd_wrapper.R
\name{dcifer_allele_freq_mean}
\alias{dcifer_allele_freq_mean}
\title{Mean of two Dcifer allele-frequency lists}
\usage{
dcifer_allele_freq_mean(af1, af2)
}
\description{
Mean of two Dcifer allele-frequency lists
}
\keyword{internal}
