% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fws_from_vcf.R
\name{derive_gds_path}
\alias{derive_gds_path}
\title{Derive a GDS path from a VCF path}
\usage{
derive_gds_path(vcf_path, gds = NULL)
}
\arguments{
\item{vcf_path}{Input VCF path.}

\item{gds}{Optional explicit GDS path.}
}
\value{
GDS path to use.
}
\description{
Derive a GDS path from a VCF path
}
\keyword{internal}
