% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_prevfreq_naive_variantstring.R
\name{extract_component_variants}
\alias{extract_component_variants}
\title{Extract unambiguous component genotypes from variant strings}
\usage{
extract_component_variants(variant_strings)
}
\arguments{
\item{variant_strings}{Character vector of variant strings.}
}
\value{
Unique component variant strings with no heterozygous loci.
}
\description{
Extract unambiguous component genotypes from variant strings
}
\keyword{internal}
