% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/filter_to_highest_diversity_independent_snp_call.R
\name{filter_highest_diversity_snps_core}
\alias{filter_highest_diversity_snps_core}
\title{Core greedy filter on expected heterozygosity and distance}
\usage{
filter_highest_diversity_snps_core(
  snp_table_in,
  mindist_between_snps = 10000,
  only_biallelic = FALSE,
  only_informative = FALSE
)
}
\description{
Core greedy filter on expected heterozygosity and distance
}
\keyword{internal}
