% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqEstimationModel_wrapper.R
\name{format_invariant_group_output}
\alias{format_invariant_group_output}
\title{Format output for an invariant FEM group}
\usage{
format_invariant_group_output(aa_calls, groups, group)
}
\description{
Format output for an invariant FEM group
}
\keyword{internal}
