% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqEstimationModel_wrapper.R
\name{format_single_group_output}
\alias{format_single_group_output}
\title{Format FEM output for one group}
\usage{
format_single_group_output(pop_freq_list)
}
\description{
Format FEM output for one group
}
\keyword{internal}
