% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_prevfreq_naive.R
\name{generate_single_locus_prev_freq_from_multilocus_groups_presence_absence}
\alias{generate_single_locus_prev_freq_from_multilocus_groups_presence_absence}
\title{Recalculate single-locus prev/freq from multilocus calls (presence/absence)}
\usage{
generate_single_locus_prev_freq_from_multilocus_groups_presence_absence(
  multilocus_calls
)
}
\arguments{
\item{multilocus_calls}{Table with \code{group_id}, \code{specimen_name}, \code{variant}.}
}
\value{
Summarized single-locus prev/freq per group.
}
\description{
Recalculate single-locus prev/freq from multilocus calls (presence/absence)
}
\keyword{internal}
