% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_prevfreq_naive.R
\name{generate_single_locus_prev_freq_from_multilocus_groups_wsaf_prop}
\alias{generate_single_locus_prev_freq_from_multilocus_groups_wsaf_prop}
\title{Recalculate single-locus prev/freq from multilocus calls (WSAF-weighted)}
\usage{
generate_single_locus_prev_freq_from_multilocus_groups_wsaf_prop(
  multilocus_calls
)
}
\arguments{
\item{multilocus_calls}{Table with \code{group_id}, \code{specimen_name}, \code{variant}, \code{wsaf}.}
}
\value{
Summarized single-locus prev/freq per group.
}
\description{
Recalculate single-locus prev/freq from multilocus calls (WSAF-weighted)
}
\keyword{internal}
