% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_microhaps.R
\name{get_aligned_pattern_from_overlap_align}
\alias{get_aligned_pattern_from_overlap_align}
\title{Aligned pattern sequence including end gaps from an overlap alignment}
\usage{
get_aligned_pattern_from_overlap_align(pw_overlapAlign)
}
\arguments{
\item{pw_overlapAlign}{A \code{\link[pwalign:pairwiseAlignment]{pwalign::pairwiseAlignment()}} overlap result.}
}
\value{
Character string of the pattern with terminal gaps relative to the
subject.
}
\description{
Aligned pattern sequence including end gaps from an overlap alignment
}
\keyword{internal}
