% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_snp_calls.R
\name{is_simple_base}
\alias{is_simple_base}
\title{TRUE if \code{x} is a single canonical DNA base (a simple-SNP allele)}
\usage{
is_simple_base(x)
}
\arguments{
\item{x}{Character string.}
}
\value{
Logical.
}
\description{
TRUE if \code{x} is a single canonical DNA base (a simple-SNP allele)
}
\keyword{internal}
