% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariaem_wrapper.R
\name{load_malariaem_allele_table}
\alias{load_malariaem_allele_table}
\title{Load and format an allele table for malaria.em}
\usage{
load_malariaem_allele_table(file_path)
}
\description{
Load and format an allele table for malaria.em
}
\keyword{internal}
