% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariaem_wrapper.R
\name{load_malariaem_target_groups}
\alias{load_malariaem_target_groups}
\title{Load target groups for subsetted malaria.em runs}
\usage{
load_malariaem_target_groups(file_path)
}
\description{
Load target groups for subsetted malaria.em runs
}
\keyword{internal}
