% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slaf_from_stave_mlaf.R
\name{load_mlaf}
\alias{load_mlaf}
\title{Load multi-locus allele frequency (MLAF) data}
\usage{
load_mlaf(path)
}
\arguments{
\item{path}{Path to a TSV with columns \code{group_id}, \code{variant}, and \code{freq}.}
}
\value{
A tibble with those three columns.
}
\description{
Load multi-locus allele frequency (MLAF) data
}
\keyword{internal}
