% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLociBiallelicModel_wrapper.R
\name{make_stave}
\alias{make_stave}
\title{Group variantstring-style mutations into stave format}
\usage{
make_stave(variant)
}
\arguments{
\item{variant}{Character string of \verb{;}-separated \code{gene:position:mutation}.}
}
\value{
A stave-formatted string.
}
\description{
Group variantstring-style mutations into stave format
}
\keyword{internal}
