% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_locus_popgen_summary.R
\name{msa_method_binary}
\alias{msa_method_binary}
\title{External MSA binary expected for a given \code{msa::msa()} method}
\usage{
msa_method_binary(msa_method)
}
\arguments{
\item{msa_method}{One of \code{"Muscle"}, \code{"ClustalW"}, or \code{"ClustalOmega"}.}
}
\value{
Executable name for \code{\link[=check_external_tool]{check_external_tool()}}.
}
\description{
External MSA binary expected for a given \code{msa::msa()} method
}
\keyword{internal}
