% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_microhaps.R
\name{nucleotide_overlap_substitution_matrix}
\alias{nucleotide_overlap_substitution_matrix}
\title{Default nucleotide substitution matrix used by pileup and translate}
\usage{
nucleotide_overlap_substitution_matrix()
}
\value{
A substitution matrix from \code{\link[pwalign:nucleotideSubstitutionMatrix]{pwalign::nucleotideSubstitutionMatrix()}}.
}
\description{
Default nucleotide substitution matrix used by pileup and translate
}
\keyword{internal}
