% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_microhaps.R
\name{orient_allele_seq_to_plus_strand}
\alias{orient_allele_seq_to_plus_strand}
\title{Orient an allele sequence to the plus strand of its target}
\usage{
orient_allele_seq_to_plus_strand(seq, strand)
}
\arguments{
\item{seq}{Character allele sequence.}

\item{strand}{Target strand (\code{"+"} or \code{"-"}).}
}
\value{
A \code{DNAString}.
}
\description{
Orient an allele sequence to the plus strand of its target
}
\keyword{internal}
