% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_microhaps.R
\name{overlap_align_allele_to_ref}
\alias{overlap_align_allele_to_ref}
\title{Overlap-align an allele DNAString to a reference sequence}
\usage{
overlap_align_allele_to_ref(allele_seq, ref_seq, mat)
}
\arguments{
\item{allele_seq}{A \code{DNAString} (already oriented to the plus strand of the
target).}

\item{ref_seq}{Character reference sequence (already oriented).}

\item{mat}{Substitution matrix from \code{nucleotide_overlap_substitution_matrix()}.}
}
\value{
A pairwise alignment object.
}
\description{
Overlap-align an allele DNAString to a reference sequence
}
\keyword{internal}
