% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THEREALMcCOIL_wrapper.R
\name{prep_input_categorical}
\alias{prep_input_categorical}
\title{Format SNP calls for the McCOIL categorical model}
\usage{
prep_input_categorical(df)
}
\arguments{
\item{df}{Output of \code{read_and_preprocess_snp_call()}.}
}
\value{
A data frame (samples x sites) of scores \code{1} / \code{0} / \code{0.5} / \code{-1}.
}
\description{
Format SNP calls for the McCOIL categorical model
}
\keyword{internal}
