% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THEREALMcCOIL_wrapper.R
\name{prep_input_prop}
\alias{prep_input_prop}
\title{Format SNP calls for the McCOIL proportional model}
\usage{
prep_input_prop(df)
}
\arguments{
\item{df}{Output of \code{read_and_preprocess_snp_call()}.}
}
\value{
A list with \code{a1} and \code{a2} read-count matrices.
}
\description{
Format SNP calls for the McCOIL proportional model
}
\keyword{internal}
