% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcifer_slaf_wrapper.R
\name{prepare_dcifer_slaf_output}
\alias{prepare_dcifer_slaf_output}
\title{Convert Dcifer allele-frequency lists to a tibble}
\usage{
prepare_dcifer_slaf_output(
  allele_freqs_list,
  n_samp_per_target,
  target_name_col = "target_name",
  target_value_col = "seq"
)
}
\description{
Convert Dcifer allele-frequency lists to a tibble
}
\keyword{internal}
