% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDM_wrapper.R
\name{prepare_input_4_aa_calls}
\alias{prepare_input_4_aa_calls}
\title{Prepare amino-acid-call input for the Incomplete Data Model}
\usage{
prepare_input_4_aa_calls(aa_calls)
}
\arguments{
\item{aa_calls}{Path to amino-acid TSV.}
}
\value{
Tibble with \code{specimen_name}, \code{locus}, \code{variants}.
}
\description{
Prepare amino-acid-call input for the Incomplete Data Model
}
\keyword{internal}
