% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDM_wrapper.R
\name{prepare_input_4_allele_table}
\alias{prepare_input_4_allele_table}
\title{Prepare allele-table input for the Incomplete Data Model}
\usage{
prepare_input_4_allele_table(allele_table_input)
}
\arguments{
\item{allele_table_input}{Path to allele TSV.}
}
\value{
Tibble with \code{specimen_name}, \code{locus}, \code{variants}.
}
\description{
Prepare allele-table input for the Incomplete Data Model
}
\keyword{internal}
