% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_allele_prevalence_naive.R
\name{prev_parse_aa_calls}
\alias{prev_parse_aa_calls}
\title{Load amino acid calls for naive allele prevalence estimation}
\usage{
prev_parse_aa_calls(path)
}
\arguments{
\item{path}{Path to a TSV of amino acid calls.}
}
\value{
A tibble with columns \code{specimen_name}, \code{target_name}, and \code{variant}.
}
\description{
Load amino acid calls for naive allele prevalence estimation
}
\keyword{internal}
