% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_allele_prevalence_naive.R
\name{prev_parse_mh_calls}
\alias{prev_parse_mh_calls}
\title{Load microhaplotype calls for naive allele prevalence estimation}
\usage{
prev_parse_mh_calls(path)
}
\arguments{
\item{path}{Path to a TSV of microhaplotype genotypes.}
}
\value{
A tibble with columns including \code{specimen_name}, \code{target_name}, and
\code{variant}.
}
\description{
Load microhaplotype calls for naive allele prevalence estimation
}
\keyword{internal}
