% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slaf_from_mhaps_freqs.R
\name{process_input_loci_of_interest_per_microhaps}
\alias{process_input_loci_of_interest_per_microhaps}
\title{Read and validate translated loci-of-interest for microhaplotypes}
\usage{
process_input_loci_of_interest_per_microhaps(
  loci_of_interest_per_microhaps_fnp
)
}
\arguments{
\item{loci_of_interest_per_microhaps_fnp}{Path to a TSV with columns
\code{target_name}, \code{gene_id}, \code{aa_position}, \code{seq}, and \code{aa}.}
}
\value{
A tibble of translated loci.
}
\description{
Read and validate translated loci-of-interest for microhaplotypes
}
\keyword{internal}
