% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slaf_from_mhaps_freqs.R
\name{process_input_mhaps_slaf}
\alias{process_input_mhaps_slaf}
\title{Read and validate microhaplotype allele frequency tables}
\usage{
process_input_mhaps_slaf(mhaps_slaf_fnp)
}
\arguments{
\item{mhaps_slaf_fnp}{Path to a TSV with columns \code{target_name}, \code{seq},
\code{freq}, and \code{sample_total}.}
}
\value{
A tibble of microhaplotype allele frequencies.
}
\description{
Read and validate microhaplotype allele frequency tables
}
\keyword{internal}
