% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THEREALMcCOIL_wrapper.R
\name{read_and_preprocess_snp_call}
\alias{read_and_preprocess_snp_call}
\title{Read SNP call data for THEREALMcCOIL}
\usage{
read_and_preprocess_snp_call(snp_calls)
}
\arguments{
\item{snp_calls}{Path to an independent, collapsed SNP call TSV.}
}
\value{
A data frame with \code{specimen_name}, \code{snp_name}, \code{seq_base}, \code{reads}.
}
\description{
Read SNP call data for THEREALMcCOIL
}
\keyword{internal}
