% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqEstimationModel_wrapper.R
\name{read_fem_aa_calls}
\alias{read_fem_aa_calls}
\title{Read amino-acid calls for FEM}
\usage{
read_fem_aa_calls(input_path)
}
\description{
Read amino-acid calls for FEM
}
\keyword{internal}
