% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqEstimationModel_wrapper.R
\name{read_fem_groups}
\alias{read_fem_groups}
\title{Read FEM loci-group definitions}
\usage{
read_fem_groups(groups_path)
}
\description{
Read FEM loci-group definitions
}
\keyword{internal}
