% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{read_genome_dna_string_set}
\alias{read_genome_dna_string_set}
\title{Load a FASTA with record names truncated at the first whitespace}
\usage{
read_genome_dna_string_set(path, reason)
}
\arguments{
\item{path}{Path to a FASTA file.}

\item{reason}{Passed to \code{\link[=check_suggested_pkg]{check_suggested_pkg()}} for \strong{Biostrings}.}
}
\value{
A \code{DNAStringSet} with shortened names.
}
\description{
Load a FASTA with record names truncated at the first whitespace
}
\keyword{internal}
