% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_biallelic_calls.R
\name{read_in_amino_acid_calls}
\alias{read_in_amino_acid_calls}
\title{Read and validate amino acid call tables for biallelic filtering}
\usage{
read_in_amino_acid_calls(amino_acid_calls_fnp)
}
\arguments{
\item{amino_acid_calls_fnp}{Path to a TSV with amino acid calls.}
}
\value{
A tibble of amino acid calls.
}
\description{
Read and validate amino acid call tables for biallelic filtering
}
\keyword{internal}
