% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{read_mhap_allele_table}
\alias{read_mhap_allele_table}
\title{Read a microhaplotype allele table (specimen_name, target_name, reads, seq)}
\usage{
read_mhap_allele_table(path)
}
\arguments{
\item{path}{Path to a TSV.}
}
\value{
A tibble.
}
\description{
Read a microhaplotype allele table (specimen_name, target_name, reads, seq)
}
\keyword{internal}
