% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_prevfreq_naive_variantstring.R
\name{read_mlp_vs_loci_groups}
\alias{read_mlp_vs_loci_groups}
\title{Read loci-group definitions for variantstring multilocus prev/freq}
\usage{
read_mlp_vs_loci_groups(loci_groups_path)
}
\arguments{
\item{loci_groups_path}{Path to a TSV with \code{group_id}, \code{gene_id}, \code{aa_position}.}
}
\value{
A tibble of loci groups.
}
\description{
Read loci-group definitions for variantstring multilocus prev/freq
}
\keyword{internal}
