% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{read_ref_bed_table}
\alias{read_ref_bed_table}
\title{Read and validate a panel reference BED table}
\usage{
read_ref_bed_table(path)
}
\arguments{
\item{path}{Path to a TSV with a header row.}
}
\value{
A tibble of the BED table.
}
\description{
Read and validate a panel reference BED table
}
\keyword{internal}
