% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{read_ref_bed_with_seq_table}
\alias{read_ref_bed_with_seq_table}
\title{Read a panel BED table that includes \code{ref_seq}}
\usage{
read_ref_bed_with_seq_table(path)
}
\arguments{
\item{path}{Path to a TSV with a header row.}
}
\value{
A tibble of the BED table.
}
\description{
Read a panel BED table that includes \code{ref_seq}
}
\keyword{internal}
