% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_microhaps.R
\name{ref_bed_lookup_by_target}
\alias{ref_bed_lookup_by_target}
\title{Index a ref_bed table by \code{target_name}}
\usage{
ref_bed_lookup_by_target(ref_bed)
}
\arguments{
\item{ref_bed}{Panel location table.}
}
\value{
Named list of one-row tibbles.
}
\description{
Index a ref_bed table by \code{target_name}
}
\keyword{internal}
