% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{ref_bed_required_cols}
\alias{ref_bed_required_cols}
\title{Required columns for panel reference BED tables used when adding ref_seq}
\usage{
ref_bed_required_cols()
}
\description{
Required columns for panel reference BED tables used when adding ref_seq
}
\keyword{internal}
