% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_calls_to_vcf.R
\name{revcomp}
\alias{revcomp}
\title{Reverse-complement a vector of DNA strings (handles single bases)}
\usage{
revcomp(x)
}
\arguments{
\item{x}{Character vector of DNA sequences.}
}
\value{
Character vector of reverse complements.
}
\description{
Reverse-complement a vector of DNA strings (handles single bases)
}
\keyword{internal}
