% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcifer_ibd_wrapper.R
\name{run_dcifer_bypop}
\alias{run_dcifer_bypop}
\title{Run Dcifer with population-specific allele frequencies}
\usage{
run_dcifer_bypop(
  allele_table_path,
  coi,
  specimen_metadata,
  specimen_name_col = "specimen_name",
  target_name_col = "target_name",
  target_value_col = "seq",
  use_estm = FALSE,
  ...
)
}
\description{
Run Dcifer with population-specific allele frequencies
}
\keyword{internal}
