% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcifer_ibd_wrapper.R
\name{run_ibdestm}
\alias{run_ibdestm}
\title{Run ibdEstM in parallel}
\usage{
run_ibdestm(
  dsmp,
  coi,
  afreq,
  sample_pairs = NULL,
  pval = TRUE,
  confint = FALSE,
  rnull = 0,
  alpha = 0.05,
  total_cores = NULL,
  verbose = FALSE
)
}
\description{
Run ibdEstM in parallel
}
\keyword{internal}
