% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcifer_ibd_wrapper.R
\name{run_ibdpair}
\alias{run_ibdpair}
\title{Run ibdPair in parallel}
\usage{
run_ibdpair(
  dsmp,
  coi,
  afreq,
  sample_pairs = NULL,
  pval = TRUE,
  confint = FALSE,
  rnull = 0,
  alpha = 0.05,
  nr = 1000,
  reval = NULL,
  total_cores = NULL,
  verbose = FALSE
)
}
\description{
Run ibdPair in parallel
}
\keyword{internal}
