% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDM_wrapper.R
\name{run_idm_mle_across_loci}
\alias{run_idm_mle_across_loci}
\title{Run IDM/OM MLE independently at each locus}
\usage{
run_idm_mle_across_loci(
  df,
  model = "IDM",
  lambda_initial = 1,
  eps_initial = 0.1
)
}
\arguments{
\item{df}{Formatted input from \code{prepare_input_4_allele_table()} or
\code{prepare_input_4_aa_calls()}.}

\item{model}{\code{"IDM"}, \code{"OM"}, \code{"IDM_OM"}, \code{"OM_CC"}, or \code{"IDM_OM_CC"}.}

\item{lambda_initial}{Initial lambda for the numerical solver.}

\item{eps_initial}{Initial epsilon for the numerical solver.}
}
\value{
Tibble with \code{variant} and \code{freq}.
}
\description{
Run IDM/OM MLE independently at each locus
}
\keyword{internal}
