% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariaem_wrapper.R
\name{run_malariaem_helper}
\alias{run_malariaem_helper}
\title{Run malaria.em on one allele matrix}
\usage{
run_malariaem_helper(matrix, test_size, max_size, label = NULL)
}
\description{
Run malaria.em on one allele matrix
}
\keyword{internal}
