% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{set_decompose}
\alias{set_decompose}
\title{Decompose shared and unique values between two vectors}
\usage{
set_decompose(vector_a, vector_b)
}
\arguments{
\item{vector_a}{First vector.}

\item{vector_b}{Second vector.}
}
\value{
A list with \code{only_in_vector_a}, \code{only_in_vector_b}, \code{shared}, and \code{all}.
}
\description{
Decompose shared and unique values between two vectors
}
\keyword{internal}
