% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{stop_on_validate_fails}
\alias{stop_on_validate_fails}
\title{Stop when validate::confront reports failing rules}
\usage{
stop_on_validate_fails(df, rules, data_name)
}
\arguments{
\item{df}{Data frame that was confronted.}

\item{rules}{A \code{\link[validate:validator]{validate::validator()}} object.}

\item{data_name}{Label used in the error message.}
}
\value{
Invisibly returns \code{TRUE} if all rules pass.
}
\description{
Stop when validate::confront reports failing rules
}
\keyword{internal}
