% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLociBiallelicModel_wrapper.R
\name{summarise_MLBM_results}
\alias{summarise_MLBM_results}
\title{Summarise MLBM haplotype frequencies as variant strings}
\usage{
summarise_MLBM_results(MLBM_res, MLBM_object, group_name)
}
\arguments{
\item{MLBM_res}{Result of \code{run_MultiLociBiallelicModel()}.}

\item{MLBM_object}{Result of \code{create_MultiLociBiallelicModel_input()}.}

\item{group_name}{Group id matching \code{MLBM_object$by_group_table}.}
}
\value{
Tibble with \code{group_id}, \code{variant}, \code{freq}.
}
\description{
Summarise MLBM haplotype frequencies as variant strings
}
\keyword{internal}
