% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moire_wrapper.R
\name{summarize_and_write_moire_results}
\alias{summarize_and_write_moire_results}
\title{Summarize MOIRe MCMC results and write TSV files}
\usage{
summarize_and_write_moire_results(
  moire_object,
  mcmc_results,
  coi_output,
  he_output,
  allele_freq_output,
  relatedness_output,
  effective_coi_output
)
}
\description{
Summarize MOIRe MCMC results and write TSV files
}
\keyword{internal}
