% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_microhaps.R
\name{unique_haps_from_allele_table}
\alias{unique_haps_from_allele_table}
\title{Unique haplotypes by target_name and seq}
\usage{
unique_haps_from_allele_table(allele_table)
}
\arguments{
\item{allele_table}{Allele table with \code{target_name} and \code{seq}.}
}
\value{
Distinct rows sorted by \code{target_name}.
}
\description{
Unique haplotypes by target_name and seq
}
\keyword{internal}
