% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coi_naive.R
\name{validate_allele_calls_for_coi_naive}
\alias{validate_allele_calls_for_coi_naive}
\title{Validate allele-call columns used for naive COI estimation}
\usage{
validate_allele_calls_for_coi_naive(df_alleles)
}
\arguments{
\item{df_alleles}{Allele-call data frame.}
}
\value{
Invisibly returns \code{TRUE} if valid.
}
\description{
Validate allele-call columns used for naive COI estimation
}
\keyword{internal}
