% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{validate_ref_bed_table}
\alias{validate_ref_bed_table}
\title{Validate panel reference BED columns and types}
\usage{
validate_ref_bed_table(ref_bed, data_name = "ref_bed")
}
\arguments{
\item{ref_bed}{Data frame of panel locations.}

\item{data_name}{Label used in error messages.}
}
\value{
Invisibly returns \code{TRUE} if valid.
}
\description{
Validate panel reference BED columns and types
}
\keyword{internal}
