% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{validate_ref_bed_with_seq_table}
\alias{validate_ref_bed_with_seq_table}
\title{Validate a panel BED table that includes \code{ref_seq}}
\usage{
validate_ref_bed_with_seq_table(ref_bed, data_name = "ref_bed")
}
\arguments{
\item{ref_bed}{Data frame of panel locations with sequences.}

\item{data_name}{Label used in error messages.}
}
\value{
Invisibly returns \code{TRUE} if valid.
}
\description{
Validate a panel BED table that includes \code{ref_seq}
}
\keyword{internal}
