% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcifer_ibd_wrapper.R
\name{write_dcifer_ibd_output}
\alias{write_dcifer_ibd_output}
\title{Write Dcifer IBD results}
\usage{
write_dcifer_ibd_output(dcifer_results, out_path)
}
\description{
Write Dcifer IBD results
}
\keyword{internal}
