% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDM_wrapper.R
\name{write_idm_output}
\alias{write_idm_output}
\title{Write IDM SLAF output}
\usage{
write_idm_output(res, slaf_output, allele_table = FALSE)
}
\arguments{
\item{res}{Result table from \code{run_idm_mle_across_loci()}.}

\item{slaf_output}{Output TSV path.}

\item{allele_table}{If \code{TRUE}, split \code{variant} into \code{target_name} and \code{seq}.}
}
\description{
Write IDM SLAF output
}
\keyword{internal}
