% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THEREALMcCOIL_wrapper.R
\name{write_mccoil_output}
\alias{write_mccoil_output}
\title{Write formatted McCOIL output}
\usage{
write_mccoil_output(df_formated, slaf_path, coi_path)
}
\arguments{
\item{df_formated}{List from \code{format_mccoil_output()}.}

\item{slaf_path}{SLAF TSV path (\code{variant}, \code{freq}).}

\item{coi_path}{COI TSV path (\code{specimen_name}, \code{coi}).}
}
\description{
Write formatted McCOIL output
}
\keyword{internal}
