% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpslice_wrapper.R
\name{create_snpslice_loci_group_input}
\alias{create_snpslice_loci_group_input}
\title{Read loci groups for SNP-Slice multilocus frequencies}
\usage{
create_snpslice_loci_group_input(
  loci_groups_path,
  allele_table,
  target_name_col = "target_name"
)
}
\description{
Read loci groups for SNP-Slice multilocus frequencies
}
\keyword{internal}
