% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THEREALMcCOIL_wrapper.R
\name{format_mccoil_output}
\alias{format_mccoil_output}
\title{Format McCOIL summary TSV into PGE COI and SLAF tables}
\usage{
format_mccoil_output(summary_path)
}
\arguments{
\item{summary_path}{Path to \verb{*_summary.txt} written by McCOIL.}
}
\value{
A list with \code{slaf} and \code{coi} tibbles.
}
\description{
Format McCOIL summary TSV into PGE COI and SLAF tables
}
\keyword{internal}
