% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fws_from_vcf.R
\name{gds_needs_conversion}
\alias{gds_needs_conversion}
\title{Whether a VCF should be (re)converted to GDS}
\usage{
gds_needs_conversion(vcf_path, gds_path, force = FALSE)
}
\arguments{
\item{vcf_path}{Input VCF path.}

\item{gds_path}{GDS path.}

\item{force}{If \code{TRUE}, always convert.}
}
\value{
Logical.
}
\description{
Whether a VCF should be (re)converted to GDS
}
\keyword{internal}
