% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGEcore-package.R
\docType{package}
\name{PGEcore-package}
\alias{PGEcore}
\alias{PGEcore-package}
\title{PGEcore: Shared PlasmoGenEpi Bioinformatics Scripts and Tool Wrappers}
\description{
Shared R functions and command-line interfaces used across PlasmoGenEpi Nextflow and WDL pipelines. Provides bespoke analysis helpers and thin wrappers around specialised malaria genomics R packages. PGEcore does not install or distribute the external software or specialised R packages that individual wrappers call; those must be installed separately and available at runtime.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PlasmoGenEpi/PGEcore}
  \item Report bugs at \url{https://github.com/PlasmoGenEpi/PGEcore/issues}
}

}
\author{
\strong{Maintainer}: Kathryn Murie \email{kathryn.murie@ucsf.edu} (\href{https://orcid.org/0009-0008-9612-2016}{ORCID})

Authors:
\itemize{
  \item Alfred Hubbard (\href{https://orcid.org/0000-0003-4917-7272}{ORCID})
  \item Nicholas Hathaway (\href{https://orcid.org/0000-0001-9639-2894}{ORCID})
  \item Shazia Ruybal Pesantez (\href{https://orcid.org/0000-0002-0495-179X}{ORCID})
  \item Jorge Eduardo Amaya-Romero
  \item Max Murphy (\href{https://orcid.org/0000-0003-0332-4388}{ORCID})
  \item Robert Verity (\href{https://orcid.org/0000-0002-3902-8567}{ORCID})
  \item Zoe Gotthold (\href{https://orcid.org/0009-0002-7092-2775}{ORCID})
  \item Bing Guo (\href{https://orcid.org/0000-0002-3998-5981}{ORCID})
  \item Alfred Simkin (\href{https://orcid.org/0000-0001-9408-0053}{ORCID})
  \item Bryan Greenhouse (\href{https://orcid.org/0000-0003-0287-9111}{ORCID})
  \item Amy Wesolowski (\href{https://orcid.org/0000-0001-6320-3575}{ORCID})
}

Other contributors:
\itemize{
  \item PlasmoGenEpi [copyright holder]
}

}
\keyword{internal}
