% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpslice_wrapper.R
\name{prepare_snpslice_af_output}
\alias{prepare_snpslice_af_output}
\title{Format SNP-Slice allele frequencies with variantstring names}
\usage{
prepare_snpslice_af_output(snpslice_res, loci_groups, estimator)
}
\description{
Format SNP-Slice allele frequencies with variantstring names
}
\keyword{internal}
