% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpslice_wrapper.R
\name{prepare_snpslice_coi_output}
\alias{prepare_snpslice_coi_output}
\title{Format SNP-Slice COI estimates}
\usage{
prepare_snpslice_coi_output(snpslice_res, specimen_name_col, estimator)
}
\description{
Format SNP-Slice COI estimates
}
\keyword{internal}
