% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_prevfreq_naive.R
\name{read_mlp_naive_aa_table}
\alias{read_mlp_naive_aa_table}
\title{Read amino acid calls for naive multilocus prev/freq}
\usage{
read_mlp_naive_aa_table(aa_table)
}
\arguments{
\item{aa_table}{Path to a TSV of amino acid calls.}
}
\value{
A tibble including \code{n_aa} (allele count per specimen/locus).
}
\description{
Read amino acid calls for naive multilocus prev/freq
}
\keyword{internal}
