% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_prevfreq_naive_variantstring.R
\name{read_mlp_vs_aa_table}
\alias{read_mlp_vs_aa_table}
\title{Read amino acid calls for variantstring multilocus prev/freq}
\usage{
read_mlp_vs_aa_table(aa_calls)
}
\arguments{
\item{aa_calls}{Path to a TSV of amino acid calls.}
}
\value{
A tibble with \code{specimen_name}, \code{gene}, \code{pos}, \code{reads}, \code{aa}, \code{n_aa}.
}
\description{
Read amino acid calls for variantstring multilocus prev/freq
}
\keyword{internal}
