% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqEstimationModel_wrapper.R
\name{run_FreqEstimationModel}
\alias{run_FreqEstimationModel}
\title{Run FreqEstimationModel MCMC for one group}
\usage{
run_FreqEstimationModel(sample_matrix_list, COI, threads, seed)
}
\description{
Run FreqEstimationModel MCMC for one group
}
\keyword{internal}
