% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moire_wrapper.R
\name{summarize_and_write_moire_results}
\alias{summarize_and_write_moire_results}
\title{Summarize MOIRe MCMC results and write TSV files}
\usage{
summarize_and_write_moire_results(
  moire_object,
  mcmc_results,
  coi_summary_o,
  he_summary_o,
  allele_freq_summary_o,
  relatedness_summary_o,
  effective_coi_summary_o
)
}
\description{
Summarize MOIRe MCMC results and write TSV files
}
\keyword{internal}
