% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_prevfreq_naive_variantstring.R
\name{write_mlp_vs_prev}
\alias{write_mlp_vs_prev}
\title{Write a variantstring prev/freq table}
\usage{
write_mlp_vs_prev(df_prev, output)
}
\arguments{
\item{df_prev}{Data frame with \code{group_id}, \code{variant}, \code{prev}, \code{freq}, \code{sample_total}.}

\item{output}{Output TSV path.}
}
\description{
Write a variantstring prev/freq table
}
\keyword{internal}
